fix(test): refactored tests of each services, added bootstrap-test.yaml on each config client service to disable config-server during testing
