+++ apps/backend/discovery-server/src/test/java/com/bytebandit/discovery/DiscoveryApplicationIT.java
@@ -1,13 +1,11 @@
-package com.bytebandit.configserver;
+package com.bytebandit.discovery;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ConfigServerApplicationTests {
-
+class DiscoveryApplicationIT {
     @Test
     void contextLoads() {
     }
-
 }
