--- apps/backend/file-service/src/test/java/com/bytebandit/fileservice/FileServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.bytebandit.fileservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FileServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
