+++ apps/backend/file-service/src/test/resources/bootstrap-test.yaml
@@ -1,0 +1,9 @@
+spring:
+  cloud:
+    config:
+      enabled: false
+      discovery:
+        enabled: false
+eureka:
+  client:
+    enabled: false
