+++ apps/backend/gateway/src/test/java/com/bytebandit/gateway/GatewayApplicationIT.java
@@ -1,0 +1,18 @@
+package com.bytebandit.gateway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest
+@ActiveProfiles("test")
+class GatewayApplicationIT {
+
+    /**
+     * Test to check if the Spring application context loads successfully.
+     */
+    @Test
+    void contextLoads() {
+    }
+
+}
