--- apps/backend/gateway/src/test/java/com/bytebandit/gateway/GatewayApplicationTests.java
@@ -1,13 +1,0 @@
-package com.bytebandit.gateway;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GatewayApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
