+++ apps/backend/user-service/src/test/java/com/bytebandit/userservice/UserServiceApplicationIT.java
@@ -1,0 +1,13 @@
+package com.bytebandit.userservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest(classes = {UserServiceApplication.class})
+@ActiveProfiles("test")
+class UserServiceApplicationIT {
+    @Test
+    void contextLoads() {
+    }
+}
